+++ lib/clickhouse/http_transport.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require "http"
+require "httpx"
 require "json"
 
 module Clickhouse
@@ -12,10 +12,15 @@ module Clickhouse
     # @param config [Config] configuration instance
     def initialize(config)
       @config = config
-      @http_client = HTTP.persistent("#{config.scheme}://#{config.host}:#{config.port}")
-        .use(:auto_deflate)
-        .use(:auto_inflate)
-        .timeout(connect: config.connection_timeout)
+      @base_url = "#{config.scheme}://#{config.host}:#{config.port}"
+      @http_client = HTTPX.plugin(:persistent, close_on_fork: true)
+        .with(
+          timeout: {connect_timeout: config.connection_timeout},
+          pool_options: {
+            max_connections_per_origin: config.pool_size,
+            pool_timeout: config.pool_timeout
+          }
+        )
 
       @default_headers = {
         "Accept-Encoding" => "gzip",
@@ -34,11 +39,13 @@ module Clickhouse
     # @raise [QueryError] if the query fails
     def execute(sql, options = {})
       query_params = {database: @config.database}.merge(options[:params] || {})
-      response = @http_client.post("/", params: query_params, body: sql, headers: @default_headers)
+      response = @http_client.post(@base_url, params: query_params, body: sql, headers: @default_headers)
 
-      summary = JSON.parse(response.headers["X-ClickHouse-Summary"], symbolize_names: true)
+      raise QueryError, response.error.message if response.error
 
-      raise QueryError, response.body.to_s unless response.status.success?
+      summary = JSON.parse(response.headers["x-clickhouse-summary"], symbolize_names: true)
+
+      raise QueryError, response.body.to_s unless response.status == 200
 
       TransportResult.new(body: response.body, summary: summary)
     end