+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+- Replace http.rb with httpx for HTTP client, remove `Pool` class ([#3](https://github.com/kukicola/clickhouse-rb/pull/3))
+
 ## [0.3.0] - 2025-12-31
 
 - Make responses more user-friendly ([#1](https://github.com/kukicola/clickhouse-rb/pull/1))